Return Cache-Control: s-maxage=0 for responses that shouldn't be cached by CDN
