+++ src/olympia/api/middleware.py
@@ -1,7 +1,7 @@
 import re
 
 from django.conf import settings
-from django.utils.cache import get_max_age, patch_cache_control, patch_vary_headers
+from django.utils.cache import patch_vary_headers
 from django.utils.deprecation import MiddlewareMixin
 
 
@@ -19,26 +19,3 @@ class APIRequestMiddleware(MiddlewareMixin):
 
     def process_exception(self, request, exception):
         self.identify_request(request)
-
-
-class APICacheControlMiddleware:
-    def __init__(self, get_response):
-        self.get_response = get_response
-
-    def __call__(self, request):
-        response = self.get_response(request)
-        request_conditions = (
-            request.is_api
-            and request.method in ('GET', 'HEAD')
-            and 'HTTP_AUTHORIZATION' not in request.META
-            and 'disable_caching' not in request.GET
-        )
-        response_conditions = (
-            not response.cookies
-            and response.status_code >= 200
-            and response.status_code < 400
-            and get_max_age(response) is None
-        )
-        if request_conditions and response_conditions:
-            patch_cache_control(response, max_age=settings.API_CACHE_DURATION)
-        return response
